training/requirements.txt
ultralytics>=8.3.0
PyYAML>=6.0
Pillow>=10.0.0
numpy>=1.26.0
